backend/requirements.txt
django==4.2
python-environ==0.4.54
djangorestframework==3.14.0
psycopg2-binary==2.9.6
django-cors-headers==4.0.0
openai==1.12.0
dj-rest-auth==4.0.1
django-allauth==0.54.0
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
Pillow==10.2.0
google-cloud-vision==3.4.2
spacy==3.7.4
uvicorn[standard]==0.27.1